// libs/common/src/enums/error-codes.general.ts
// libs/common/enums/error-codes.general.ts
export enum GeneralErrorCodes {
    BAD_REQUEST = 'BAD_REQUEST',
    UNAUTHORIZED = 'UNAUTHORIZED',
    FORBIDDEN = 'FORBIDDEN',
    NOT_FOUND = 'NOT_FOUND',
    CONFLICT = 'CONFLICT',
    VALIDATION_ERROR = 'VALIDATION_ERROR',
    UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY',
    TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
    REQUEST_TIMEOUT = 'REQUEST_TIMEOUT',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    BAD_GATEWAY = 'BAD_GATEWAY',
    SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
    GATEWAY_TIMEOUT = 'GATEWAY_TIMEOUT',
}

export enum DatabaseErrorCodes {
    DATABASE_CONNECTION_FAILED = 'DATABASE_CONNECTION_FAILED',
    DUPLICATE_ENTRY = 'DUPLICATE_ENTRY',
    FOREIGN_KEY_CONSTRAINT = 'FOREIGN_KEY_CONSTRAINT',
    TRANSACTION_FAILED = 'TRANSACTION_FAILED',
    QUERY_TIMEOUT = 'QUERY_TIMEOUT',
}

export enum ValidationErrorCodes {
    INVALID_EMAIL_FORMAT = 'INVALID_EMAIL_FORMAT',
    INVALID_PHONE_FORMAT = 'INVALID_PHONE_FORMAT',
    INVALID_URL_FORMAT = 'INVALID_URL_FORMAT',
    INVALID_DATE_RANGE = 'INVALID_DATE_RANGE',
    REQUIRED_FIELD_MISSING = 'REQUIRED_FIELD_MISSING',
    VALUE_TOO_SHORT = 'VALUE_TOO_SHORT',
    VALUE_TOO_LONG = 'VALUE_TOO_LONG',
    INVALID_ENUM_VALUE = 'INVALID_ENUM_VALUE',
    INVALID_UUID = 'INVALID_UUID',
}

export enum ExternalServiceErrorCodes {
    EMAIL_SERVICE_UNAVAILABLE = 'EMAIL_SERVICE_UNAVAILABLE',
    EMAIL_SEND_FAILED = 'EMAIL_SEND_FAILED',
    SMS_SERVICE_UNAVAILABLE = 'SMS_SERVICE_UNAVAILABLE',
    SMS_SEND_FAILED = 'SMS_SEND_FAILED',
    STORAGE_SERVICE_UNAVAILABLE = 'STORAGE_SERVICE_UNAVAILABLE',
    THIRD_PARTY_API_ERROR = 'THIRD_PARTY_API_ERROR',
}