// libs/common/src/enums/error-codes.auth.ts
// libs/common/enums/error-codes.auth.ts
export enum AuthErrorCodes {
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  PHONE_ALREADY_EXISTS = "PHONE_ALREADY_EXISTS",
  WEAK_PASSWORD = "WEAK_PASSWORD",
  PASSWORD_MISMATCH = "PASSWORD_MISMATCH",
  EMAIL_NOT_VERIFIED = "EMAIL_NOT_VERIFIED",
  EMAIL_ALREADY_VERIFIED = "EMAIL_ALREADY_VERIFIED",
  VERIFICATION_TOKEN_INVALID = "VERIFICATION_TOKEN_INVALID",
  VERIFICATION_TOKEN_EXPIRED = "VERIFICATION_TOKEN_EXPIRED",
  VERIFICATION_ALREADY_SENT = "VERIFICATION_ALREADY_SENT",
  SESSION_EXPIRED = "SESSION_EXPIRED",
  SESSION_INVALID = "SESSION_INVALID",
  SESSION_NOT_FOUND = "SESSION_NOT_FOUND",
  INVALID_TOKEN = "INVALID_TOKEN",
  TOKEN_EXPIRED = "TOKEN_EXPIRED",
  TOKEN_ALREADY_USED = "TOKEN_ALREADY_USED",
  RESET_TOKEN_INVALID = "RESET_TOKEN_INVALID",
  RESET_TOKEN_EXPIRED = "RESET_TOKEN_EXPIRED",
  PASSWORD_RECENTLY_CHANGED = "PASSWORD_RECENTLY_CHANGED",
  ACCOUNT_SUSPENDED = "ACCOUNT_SUSPENDED",
  ACCOUNT_DELETED = "ACCOUNT_DELETED",
  ACCOUNT_LOCKED = "ACCOUNT_LOCKED",
  INSUFFICIENT_PERMISSIONS = "INSUFFICIENT_PERMISSIONS",
  ROLE_NOT_AUTHORIZED = "ROLE_NOT_AUTHORIZED",
}

export enum OAuthErrorCodes {
  OAUTH_INIT_FAILED = "OAUTH_INIT_FAILED",
  OAUTH_CALLBACK_FAILED = "OAUTH_CALLBACK_FAILED",
  OAUTH_STATE_INVALID = "OAUTH_STATE_INVALID",
  OAUTH_STATE_EXPIRED = "OAUTH_STATE_EXPIRED",
  OAUTH_CODE_INVALID = "OAUTH_CODE_INVALID",
  OAUTH_TOKEN_EXCHANGE_FAILED = "OAUTH_TOKEN_EXCHANGE_FAILED",
  GOOGLE_AUTH_FAILED = "GOOGLE_AUTH_FAILED",
  GOOGLE_TOKEN_INVALID = "GOOGLE_TOKEN_INVALID",
  GOOGLE_ID_TOKEN_INVALID = "GOOGLE_ID_TOKEN_INVALID",
  GITHUB_AUTH_FAILED = "GITHUB_AUTH_FAILED",
  GITHUB_TOKEN_INVALID = "GITHUB_TOKEN_INVALID",
  OAUTH_CONFIG_INCOMPLETE = "OAUTH_CONFIG_INCOMPLETE",
  OAUTH_PROVIDER_NOT_CONFIGURED = "OAUTH_PROVIDER_NOT_CONFIGURED",
  OAUTH_EMAIL_DOMAIN_BLOCKED = "OAUTH_EMAIL_DOMAIN_BLOCKED",
  OAUTH_ROLE_MISMATCH = "OAUTH_ROLE_MISMATCH",
  OAUTH_ACCOUNT_LINKING_FAILED = "OAUTH_ACCOUNT_LINKING_FAILED",
  OAUTH_MISSING_PARAMETERS = "OAUTH_MISSING_PARAMETERS",
  OAUTH_INVALID_ROLE = "OAUTH_INVALID_ROLE",
}

export enum RateLimitErrorCodes {
  RATE_LIMIT_EXCEEDED = "RATE_LIMIT_EXCEEDED",
  TOO_MANY_LOGIN_ATTEMPTS = "TOO_MANY_LOGIN_ATTEMPTS",
  TOO_MANY_VERIFICATION_REQUESTS = "TOO_MANY_VERIFICATION_REQUESTS",
  TOO_MANY_PASSWORD_RESET_REQUESTS = "TOO_MANY_PASSWORD_RESET_REQUESTS",
  API_QUOTA_EXCEEDED = "API_QUOTA_EXCEEDED",
  APPROACHING_DAILY_LIMIT = "APPROACHING_DAILY_LIMIT",
}