// libs/common/src/enums/error-codes.domain.ts
export enum UserErrorCodes {
  USER_NOT_FOUND = "USER_NOT_FOUND",
  PROFILE_NOT_FOUND = "PROFILE_NOT_FOUND",
  PROFILE_INCOMPLETE = "PROFILE_INCOMPLETE",
  PROFILE_NOT_CREATED = "PROFILE_NOT_CREATED",
  PROFILE_ALREADY_EXISTS = "PROFILE_ALREADY_EXISTS",
  INVALID_PROFILE_DATA = "INVALID_PROFILE_DATA",
  PROFILE_UPDATE_FAILED = "PROFILE_UPDATE_FAILED",
}

export enum UploadErrorCodes {
  FILE_TOO_LARGE = "FILE_TOO_LARGE",
  INVALID_FILE_TYPE = "INVALID_FILE_TYPE",
  FILE_UPLOAD_FAILED = "FILE_UPLOAD_FAILED",
  FILE_NOT_FOUND = "FILE_NOT_FOUND",
  FILE_DELETE_FAILED = "FILE_DELETE_FAILED",
  INVALID_FILE_FORMAT = "INVALID_FILE_FORMAT",
  FILE_CORRUPTED = "FILE_CORRUPTED",
  STORAGE_QUOTA_EXCEEDED = "STORAGE_QUOTA_EXCEEDED",
}

export enum JobErrorCodes {
  JOB_NOT_FOUND = "JOB_NOT_FOUND",
  JOB_ALREADY_CLOSED = "JOB_ALREADY_CLOSED",
  JOB_ALREADY_EXPIRED = "JOB_ALREADY_EXPIRED",
  JOB_NOT_EDITABLE = "JOB_NOT_EDITABLE",
  JOB_NOT_DELETABLE = "JOB_NOT_DELETABLE",
  JOB_CREATION_FAILED = "JOB_CREATION_FAILED",
  INVALID_JOB_STATUS = "INVALID_JOB_STATUS",
  JOB_LIMIT_REACHED = "JOB_LIMIT_REACHED",
  DRAFT_SAVE_FAILED = "DRAFT_SAVE_FAILED",
  PROFILE_COMPLETION_REQUIRED = "PROFILE_COMPLETION_REQUIRED",
  COMPANY_NOT_VERIFIED = "COMPANY_NOT_VERIFIED",
  VERIFICATION_PENDING = "VERIFICATION_PENDING",
}

export enum ApplicationErrorCodes {
  APPLICATION_NOT_FOUND = "APPLICATION_NOT_FOUND",
  APPLICATION_SUBMISSION_FAILED = "APPLICATION_SUBMISSION_FAILED",
  NO_APPLICATIONS_FOUND = "NO_APPLICATIONS_FOUND",
  ALREADY_APPLIED = "ALREADY_APPLIED",
  JOB_NOT_ACCEPTING_APPLICATIONS = "JOB_NOT_ACCEPTING_APPLICATIONS",
  JOB_EXPIRED = "JOB_EXPIRED",
  JOB_CLOSED = "JOB_CLOSED",
  CANNOT_APPLY_OWN_JOB = "CANNOT_APPLY_OWN_JOB",
  RESUME_REQUIRED = "RESUME_REQUIRED",
  INVALID_RESUME = "INVALID_RESUME",
  RESUME_NOT_FOUND = "RESUME_NOT_FOUND",
  SCREENING_ANSWERS_REQUIRED = "SCREENING_ANSWERS_REQUIRED",
  SCREENING_ANSWERS_INVALID = "SCREENING_ANSWERS_INVALID",
  SCREENING_ANSWER_TOO_SHORT = "SCREENING_ANSWER_TOO_SHORT",
  REQUIRED_QUESTION_MISSING = "REQUIRED_QUESTION_MISSING",
  INVALID_QUESTION_ID = "INVALID_QUESTION_ID",
  PROFILE_COMPLETION_TOO_LOW = "PROFILE_COMPLETION_TOO_LOW",
  CANNOT_WITHDRAW = "CANNOT_WITHDRAW",
  ALREADY_WITHDRAWN = "ALREADY_WITHDRAWN",
  CANNOT_WITHDRAW_REJECTED = "CANNOT_WITHDRAW_REJECTED",
  INVALID_STATUS_TRANSITION = "INVALID_STATUS_TRANSITION",
  WITHDRAWN_APPLICATION_LOCKED = "WITHDRAWN_APPLICATION_LOCKED",
  REJECTED_APPLICATION_LOCKED = "REJECTED_APPLICATION_LOCKED",
  STATUS_UPDATE_FAILED = "STATUS_UPDATE_FAILED",
  NOT_YOUR_APPLICATION = "NOT_YOUR_APPLICATION",
  NOT_YOUR_JOB = "NOT_YOUR_JOB",
  EMPLOYER_PROFILE_REQUIRED = "EMPLOYER_PROFILE_REQUIRED",
  AI_SCORE_CALCULATION_FAILED = "AI_SCORE_CALCULATION_FAILED",
  AI_SCORE_NOT_AVAILABLE = "AI_SCORE_NOT_AVAILABLE",
  PROFILE_INCOMPLETE_FOR_SCORING = "PROFILE_INCOMPLETE_FOR_SCORING",
  AI_SCORE_TIMEOUT = "AI_SCORE_TIMEOUT",
  APPLICATION_CLOSED = "APPLICATION_CLOSED",
  APPLICATION_EXPIRED = "APPLICATION_EXPIRED",
  INVALID_APPLICATION_STATUS = "INVALID_APPLICATION_STATUS",
  COVER_LETTER_REQUIRED = "COVER_LETTER_REQUIRED",
  APPLICATION_LIMIT_REACHED = "APPLICATION_LIMIT_REACHED",
}

export enum VerificationErrorCodes {
  VERIFICATION_NOT_FOUND = "VERIFICATION_NOT_FOUND",
  EMPLOYER_PROFILE_NOT_FOUND = "EMPLOYER_PROFILE_NOT_FOUND",
  ALREADY_VERIFIED = "ALREADY_VERIFIED",
  VERIFICATION_PENDING = "VERIFICATION_PENDING",
  VERIFICATION_UNDER_REVIEW = "VERIFICATION_UNDER_REVIEW",
  VERIFICATION_REJECTED = "VERIFICATION_REJECTED",
  INVALID_VERIFICATION_DOCUMENT = "INVALID_VERIFICATION_DOCUMENT",
  DOCUMENT_UPLOAD_FAILED = "DOCUMENT_UPLOAD_FAILED",
  DOCUMENT_EXPIRED = "DOCUMENT_EXPIRED",
  DOCUMENT_TOO_LARGE = "DOCUMENT_TOO_LARGE",
  UNSUPPORTED_DOCUMENT_TYPE = "UNSUPPORTED_DOCUMENT_TYPE",
  VERIFICATION_REQUIRED = "VERIFICATION_REQUIRED",
  PROFILE_INCOMPLETE_FOR_VERIFICATION = "PROFILE_INCOMPLETE_FOR_VERIFICATION",
  RESUBMISSION_LIMIT_REACHED = "RESUBMISSION_LIMIT_REACHED",
  CANNOT_RESUBMIT_YET = "CANNOT_RESUBMIT_YET",
  VERIFICATION_SUBMISSION_FAILED = "VERIFICATION_SUBMISSION_FAILED",
  VERIFICATION_UPDATE_FAILED = "VERIFICATION_UPDATE_FAILED",
}

export enum LocationErrorCodes {
  LOCATION_NOT_FOUND = "LOCATION_NOT_FOUND",
  INVALID_LOCATION_DATA = "INVALID_LOCATION_DATA",
  LOCATION_SERVICE_UNAVAILABLE = "LOCATION_SERVICE_UNAVAILABLE",
  LOCATION_SERVICE_ERROR = "LOCATION_SERVICE_ERROR",
  LOCATION_API_UNAVAILABLE = "LOCATION_API_UNAVAILABLE",
  GEOCODING_FAILED = "GEOCODING_FAILED",
}

export enum NotificationErrorCodes {
  NOTIFICATION_NOT_FOUND = "NOTIFICATION_NOT_FOUND",
  NOTIFICATION_SEND_FAILED = "NOTIFICATION_SEND_FAILED",
  NOTIFICATION_FAILED = "NOTIFICATION_FAILED",
  INVALID_NOTIFICATION_CHANNEL = "INVALID_NOTIFICATION_CHANNEL",
  NOTIFICATION_PREFERENCES_NOT_FOUND = "NOTIFICATION_PREFERENCES_NOT_FOUND",
}

export enum MessageErrorCodes {
  MESSAGE_NOT_FOUND = "MESSAGE_NOT_FOUND",
  CONVERSATION_NOT_FOUND = "CONVERSATION_NOT_FOUND",
  CANNOT_MESSAGE_SELF = "CANNOT_MESSAGE_SELF",
  MESSAGE_SEND_FAILED = "MESSAGE_SEND_FAILED",
  BLOCKED_USER = "BLOCKED_USER",
  CONVERSATION_CLOSED = "CONVERSATION_CLOSED",
  INVALID_MESSAGE_TYPE = "INVALID_MESSAGE_TYPE",
  MESSAGE_TOO_LONG = "MESSAGE_TOO_LONG",
  ENCRYPTION_FAILED = "ENCRYPTION_FAILED",
  DECRYPTION_FAILED = "DECRYPTION_FAILED",
  CONVERSATION_ACCESS_DENIED = "CONVERSATION_ACCESS_DENIED",
  ROLE_MISMATCH = "ROLE_MISMATCH",
}

export enum AIErrorCodes {
  AI_SERVICE_UNAVAILABLE = "AI_SERVICE_UNAVAILABLE",
  AI_GENERATION_FAILED = "AI_GENERATION_FAILED",
  AI_QUOTA_EXCEEDED = "AI_QUOTA_EXCEEDED",
  INVALID_AI_PROMPT = "INVALID_AI_PROMPT",
  AI_TIMEOUT = "AI_TIMEOUT",
}

export enum PaymentErrorCodes {
  PAYMENT_FAILED = "PAYMENT_FAILED",
  INVALID_PAYMENT_METHOD = "INVALID_PAYMENT_METHOD",
  PAYMENT_REQUIRED = "PAYMENT_REQUIRED",
  SUBSCRIPTION_EXPIRED = "SUBSCRIPTION_EXPIRED",
  INSUFFICIENT_CREDITS = "INSUFFICIENT_CREDITS",
}

export enum SavedJobErrorCodes {
  SAVED_JOB_NOT_FOUND = "SAVED_JOB_NOT_FOUND",
  JOB_ALREADY_SAVED = "JOB_ALREADY_SAVED",
  CANNOT_SAVE_OWN_JOB = "CANNOT_SAVE_OWN_JOB",
  SAVED_JOBS_LIMIT_REACHED = "SAVED_JOBS_LIMIT_REACHED",
}

export enum PublicProfileErrorCodes {
  SEEKER_PROFILE_NOT_PUBLIC = "SEEKER_PROFILE_NOT_PUBLIC",
  COMPANY_PROFILE_NOT_PUBLIC = "COMPANY_PROFILE_NOT_PUBLIC",
}

export enum CompanyErrorCodes {
  COMPANY_NOT_FOUND = "COMPANY_NOT_FOUND",
  COMPANY_SLUG_TAKEN = "COMPANY_SLUG_TAKEN",
  INVALID_COMPANY_DATA = "INVALID_COMPANY_DATA",
}

export enum SettingsErrorCodes {
  SETTINGS_NOT_FOUND = "SETTINGS_NOT_FOUND",
  EMAIL_UNCHANGED = "EMAIL_UNCHANGED",
  PASSWORD_UNCHANGED = "PASSWORD_UNCHANGED",
  INVALID_PASSWORD = "INVALID_PASSWORD",
}

export enum LocationPreferenceErrorCodes {
  MAX_LOCATIONS_REACHED = "MAX_LOCATIONS_REACHED",
  LOCATION_ALREADY_ADDED = "LOCATION_ALREADY_ADDED",
  LOCATION_PREFERENCE_NOT_FOUND = "LOCATION_PREFERENCE_NOT_FOUND",
}